+++ app/package.json
@@ -4,13 +4,14 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon -r esm index.js",
+    "start": "nodemon -r esm server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^8.2.0",
     "esm": "^3.2.25",
     "express": "^4.17.1",
     "express-winston": "^4.0.5",
@@ -21,5 +22,13 @@
     "winston-daily-rotate-file": "^4.5.0",
     "winston-elasticsearch": "^0.10.0",
     "winston-mongodb": "^5.0.5"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "babel-jest": "^26.3.0",
+    "jest": "^26.4.2",
+    "superagent": "^6.1.0",
+    "supertest": "^5.0.0"
   }
 }
