+++ app/server.js
@@ -1,0 +1,7 @@
+import app from "./app";
+
+const port = 3000;
+
+app.listen(port, () => {
+    console.log('app is running on port', port);
+});
